requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-admin-soft-dashboard==1.0.11
django-cors-headers==4.2.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
payos==0.1.8
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
